pyproject.toml
[tool.poetry]
name = "jelkob"
description = "Main website for accessing and managing Jelka FMF"
license = "GPL-3.0-or-later"
version = "0.0.0"

homepage = "https://github.com/Jelka-FMF/Jelkob"
repository = "https://github.com/Jelka-FMF/Jelkob"
authors = ["Programerski klub FMF"]

[tool.poetry.dependencies]
python = "^3.10"
daphne = "^4.1.2"
django = "^5.1.2"
djangorestframework = "^3.15.2"
django-solo = "^2.4.0"
django-filter = "^24.3"
django-simple-history = "^3.7.0"
django-templates-macros = "^0.3"
django-eventstream = {extras = ["drf"], version = "^5.3.1"}

[tool.poetry.group.dev.dependencies]
ruff = "^0.7.1"

[tool.ruff]
line-length = 110

[tool.ruff.lint]
select = ["F", "E", "W", "UP", "DJ", "B", "G", "RUF"]
ignore = ["E203", "E501", "RUF012"]

[tool.poetry-dynamic-versioning]
enable = true

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
build-backend = "poetry.core.masonry.api"
